app/requirements.txt
asgiref==3.8.1
attrs==23.2.0
brazilcep==6.3.1
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.0.7
djangorestframework==3.15.2
idna==3.7
iniconfig==2.0.0
isodate==0.6.1
lxml==5.2.2
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pytest==8.2.2
pytest-django==4.8.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
six==1.16.0
sqlparse==0.5.0
urllib3==2.2.2
zeep==4.2.1
